config: Cast price_stop_level to (float)

+++ config/EURUSD_H4.h
@@ -24,7 +24,7 @@ struct Stg_RSI_Params_H4 : StgParams {
     signal_close_method = 0;
     signal_close_level = (float)30;
     price_stop_method = 0;
-    price_stop_level = 15;
+    price_stop_level = (float)15;
     tick_filter_method = 1;
     max_spread = 0;
   }
